refactor: result型のりファクタ
